% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custompackage-package.R
\docType{package}
\name{custompackage-package}
\alias{custompackage}
\alias{custompackage-package}
\title{custompackage: Linear discriminant function analysis}
\description{
The functions included here allow you to perform linear discriminant function analysis.
}
\author{
\strong{Maintainer}: Samantha Smith \email{samksmith@utexas.edu} (0000-0003-4491-1028)

Authors:
\itemize{
  \item Sunishka Thakur \email{sunishka@utexas.edu}
}

}
\keyword{internal}
