% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknife.R
\name{jackknife}
\alias{jackknife}
\title{jackknife approach to evaluate an LDA}
\usage{
jackknife(df)
}
\arguments{
\item{one}{line for each parameter in our function}
}
\description{
jackknife approach to evaluate an LDA
}
\keyword{aid}
\keyword{functions}
\keyword{in}
\keyword{searching}
\keyword{to}
