% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_comparison.R
\name{lda_comparison}
\alias{lda_comparison}
\title{comparison tables from LDA}
\usage{
lda_comparison(testing_df, predictions)
}
\arguments{
\item{testing_df}{dataframe with column "group" and LDA to be run on all other variables}

\item{predictions}{variable in which results of lda() and predict() (from MASS package) are stored}
}
\description{
This function outputs a table that compares the original groups that the samples are assigned to, and the group that have been assigned by the LDA.
}
\examples{
lda = lda(data = training, group ~ .)
lda_comparison(training, predictions)
}
\keyword{LDA}
\keyword{comparison}
\keyword{tables}
