% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_graphs.R
\name{lda_graphs}
\alias{lda_graphs}
\title{graphical results from LDA}
\usage{
lda_graphs(df)
}
\arguments{
\item{dataframe}{with column "group" and LDA to be run on all other variables}
}
\description{
This function performs an LDA on a dataset and outputs a scatterplot of LD1 vs LD2, and loading histograms (using ggplot, similar to outputs of MASS::ldahist) for each discriminant function
}
\keyword{LDA}
\keyword{histogram}
\keyword{plot}
