% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_process.R
\name{lda_process}
\alias{lda_process}
\title{pre processing testing and training datasets}
\usage{
lda_process(df, partition = 0.8)
}
\arguments{
\item{df}{dataframe with column "group" and LDA to be run on all other variables}

\item{partition}{divides your data up to the percent specified as training dataset, and 1- percent specified as testing dataset. For a 80-20 splitfor example, enter 0.8}
}
\description{
this function performs an LDA on a dataframe with groups in column "group". It performs multiple LDAs on the dataframe by removing one observation each time. The LDA is validated by comparing the results of these LDAs against one another. A graph is outputted from this function that shows how often each sample was sorted into a single group.
}
\examples{
lda = lda(data = training, group ~ .)
predictions = predict(lda)
lda_comparison(data, predictions)
}
\keyword{preProcess}
\keyword{testing}
\keyword{training}
