% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top5vars.R
\name{top5vars}
\alias{top5vars}
\title{top 5 variables loading on to discriminant functions}
\usage{
top5vars(df)
}
\arguments{
\item{dataframe}{with column "group" and LDA to be run on all other variables}
}
\description{
this function performs an LDA on a dataset with three groups and then outputs the top 5 variables loading on to discriminant functions 1 (LD1) and 2 (LD2)
}
\keyword{LDA}
\keyword{loadings}
\keyword{scores}
