% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lda_comparison.R
\name{lda_comparison}
\alias{lda_comparison}
\title{tabular results from LDA}
\usage{
lda_comparison(df)
}
\arguments{
\item{df}{data frame to run LDA on. Must have column named "group"}
}
\description{
tabular results from LDA
}
\keyword{aid}
\keyword{functions}
\keyword{in}
\keyword{searching}
\keyword{to}
